requirements.txt
asgiref==3.2.5
Django==3.0.7
django-cors-headers==3.5.0
djangorestframework==3.11.0
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.1
